renamed lib to src
