+++ pages/_document.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Document, { Head, Main, NextScript } from "next/document";
 import { ServerStyleSheets } from "@material-ui/styles";
-import theme from "../lib/theme";
+import theme from "../src/theme";
 
 class MyDocument extends Document {
   render() {
